add a custom hook
